fix: rename exceptions -> exception
