+++ bayes_opt/exception.py
